expose dynamically loaded cuda driver api
